+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add `onRecover` feature to `RetryConfiguration`. It's executed when all retry has failed. ([#100](https://github.com/batch-dart/batch.dart/issues/100))
 - Supported to output log errors during parallel processing. ([#99](https://github.com/batch-dart/batch.dart/issues/99))
 - Added `onLoadArgs` callback in the `BatchApplication`. ([#102](https://github.com/batch-dart/batch.dart/issues/102))
+- Enhanced customizability of `LogPrinter`. ([#107](https://github.com/batch-dart/batch.dart/issues/107))
 
 ## v0.7.1